requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
pydantic==2.12.4
pydantic-settings==2.12.0
SQLAlchemy==2.0.44
alembic==1.17.2
asyncpg==0.30.0
redis==7.1.0
python-dotenv==1.2.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.20
email-validator==2.3.0
langchain==1.0.8
langchain-groq==1.0.1
langgraph==1.0.3
groq==0.36.0
geopy==2.4.1
pytest==9.0.1
pytest-asyncio==1.3.0
requests==2.32.5